fastapi_service/requirements.txt
fastapi==0.115.12
uvicorn==0.23.2

kafka-python==2.0.2
redis==6.0.0

SQLAlchemy==2.0.40
asyncpg==0.29.0

python-dotenv==1.1.0
loguru==0.7.3
